// src/generated/model/oppgaveElementHendelsetype.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type OppgaveElementHendelsetype = (typeof OppgaveElementHendelsetype)[keyof typeof OppgaveElementHendelsetype];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OppgaveElementHendelsetype = {
    dokumentasjonEtterspurt: "dokumentasjonEtterspurt",
    dokumentasjonkrav: "dokumentasjonkrav",
    soknad: "soknad",
    bruker: "bruker",
} as const;